fix CurrentUserService and Authentication

+++ DomainNew/Models/Review.cs
@@ -4,8 +4,9 @@ namespace DomainNew.Models
 {
     public class Review
     {
+
         public long Id { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
         public int PetsitterId { get; set; }
         public Petsitter Petsitter { get; set; }
